pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vectora"
version = "5.0.0"
description = "Vectora - Herramienta Profesional de Gestión de PDF"
authors = [
    {name = "Carlo", email = "carlo@vectora.app"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6==6.10.1",
    "PyPDF2==3.0.1",
    "pikepdf==10.1.0",
    "pdf2docx==0.5.8",
    "pdf2image==1.17.0",
    "PyMuPDF==1.26.7",
    "pillow==12.1.0",
    "opencv-python==4.12.0.88",
    "pytesseract==0.3.13",
    "python-dotenv==1.2.1",
    "watchdog==6.0.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --strict-markers --tb=short --disable-warnings --cov=backend --cov=utils --cov=ui --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests for individual functions",
    "integration: Integration tests for workflows",
    "slow: Tests that take significant time",
    "pdf: Tests that work with PDF files",
]

[tool.coverage.run]
source = ["backend", "utils", "ui"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/dist/*",
    "*/build/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "PySide6.*",
    "PyPDF2.*",
    "pikepdf.*",
    "pdf2docx.*",
    "pdf2image.*",
    "fitz.*",
    "cv2.*",
    "pytesseract.*",
    "watchdog.*",
    "pytest.*"
]
ignore_missing_imports = true

[tool.pylint.messages_control]
disable = [
    "C0111", # missing-docstring
    "C0103", # invalid-name
    "C0330", # bad-continuation
    "C0326", # bad-whitespace
    "W0621", # redefined-outer-name (useful for pytest fixtures)
]
max-line-length = 100

[tool.pylint.format]
max-line-length = 100
